feat(HW-549): Receive screen refactoring
