+++ src/screens/HomeStack/ReceiveStack/index.ts
@@ -1,0 +1,1 @@
+export * from './receive-screen';
